src/requirements.txt
# stdlib only; grpcio required
grpcio>=1.60
protobuf>=4.25
